notification_backend/requirements.txt
aiokafka==0.12.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
async-firebase==4.0.0
async-timeout==5.0.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
dotenv==0.9.9
fast-depends==2.4.12
fastapi==0.115.12
faststream==0.5.41
google-auth==2.38.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-logging-loki==0.3.1
requests==2.32.3
rfc3339==6.2
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
